chore: next.config.(js|mjs) -> move next.config.js to next.config.mjs (ecma script module). Renamed src/configs from .ts to .js
