+++ package.json
@@ -2,6 +2,7 @@
 	"name": "sharebook-frontend",
 	"version": "0.1.0",
 	"private": true,
+	"type": "module",
 	"scripts": {
 		"dev": "next dev",
 		"dev-docker": "docker-compose -f docker-compose.dev.yml up",
