+++ src/configs/index.js
@@ -1,9 +1,6 @@
-interface ConfigType {
-	ApiUrl: string;
-	viaCepUrl: string;
-}
+//Needs is .js file because is used into next.config.mjs
 
-const config: ConfigType = {
+const config = {
 	ApiUrl: process.env.NEXT_PUBLIC_SHAREBOOK_API ?? 'https://sharebook.com.br/api/',
 	viaCepUrl: 'https://viacep.com.br/'
 };
